moving testConnectionCtx to Latest test (failing with new geography type) (#398)
